// tool/modules/scripts/popcap/resources/strict/property.ts
"use strict";
export enum Resources_Property {
    height = "ah",
    width = "aw",
    groups = "groups",
    position_x = "ax",
    position_y = "ay",
    cols = "cols",
    resources = "resources",
    pea_shadows = "IMAGE_PEA_SHADOWS",
    popcap_resource_id = "id",
    default_pea_shadows_property_fix = 2,
    popcap_resource_slot = "slot",
}